+++ noodles-fasta/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * fasta/indexed_reader: Add getter for index
+    (`IndexedReader::index`).
+
 ## 0.19.0 - 2023-03-03
 
 ### Changed